requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.1
cycler==0.12.1
fastapi==0.111.0
filelock==3.15.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
jax==0.4.28
jaxlib==0.4.28
kiwisolver==1.4.5
matplotlib==3.9.0
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.3
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.13.1
six==1.16.0
sounddevice==0.4.7
sympy==1.12.1
torch==2.3.1
ujson==5.10.0
pydantic==2.7.2
sqlalchemy==2.0.33
psycopg2==2.9.9
passlib==1.7.4
bcrypt==4.2.0
alembic==1.13.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
python-dotenv-vault==0.6.4
PyJWT==2.9.0
pytz==2024.2
itsdangerous==2.2.0
fastapi-mail==1.4.1